Optimize S3 Generation: FP16 Support & Enhanced Batching
- Enabled FP16 support for S3Gen (Flow + HiFiGAN) offering 2x speedup
- Fixed dtype mismatch errors in S3Gen components (HifiGAN, Flow, F0Predictor)
- Implemented robust batching logic (internal S3 batch size 8)
- Added 'test_streaming_interleaved.py' and 'test_s3_batching_audio.py' for validation
- Updated benchmark script for production readiness testing
- Streaming now supports interleaved chunks for concurrent requests

+++ src/chatterbox_vllm/models/s3gen/flow.py
@@ -312,4 +312,5 @@ class CausalMaskedDiffWithXvec(torch.nn.Module):
         )
         feat = feat[:, :, mel_len1:]
         assert feat.shape[2] == mel_len2
-        return feat.float(), None  # NOTE jrm: why are they returning None here?
+        # Keep the dtype from the model (FP16 if fp16=True, else FP32)
+        return feat, None  # NOTE jrm: why are they returning None here?
